feat(mattermost): Add emoji support
